pyproject.toml
[project]
name = "aio-sensor-intelligence"
version = "0.1.0"
description = "Physical AI at the Edge — Deploy time-series foundation models on Kubernetes with Azure IoT Operations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "paho-mqtt>=2.1.0",
    "momentfm>=1.0.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]
moirai = [
    "uni2ts",
]
chronos = [
    "chronos-forecasting",
]
notebooks = [
    "jupyter>=1.0.0",
    "matplotlib>=3.9.0",
    "pandas>=2.2.0",
]
all = [
    "aio-sensor-intelligence[dev,moirai,chronos,notebooks]",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py310"
